feat(platform): use generateNextAuthConfig from e-class-auth package
