+++ packages/auth/src/index.ts
@@ -7,7 +7,7 @@ export {AuthProviderProfileDTOSchema, type TAuthProviderProfileDTO } from './cor
 export {GetSessionDTOSchema, type TGetSessionDTO, ExtractJWTDTOSchema, type TExtractJWTDTO} from './core/dto/auth-gateway-dto';
 export { type AuthGatewayOutputPort} from './core/ports/secondary/auth-gateway-output-port';
 export { NextAuthGateway } from './infrastructure/gateway/next-auth-gateway';
-
+export { generateNextAuthConfig } from './infrastructure/config/next-auth.config';
 export { type isSessionAware, extractPlatformSpecificRoles } from './infrastructure/utils';
 interface TPermission {
     course: TCourseResourcePermission
